feat: move discovery stats strip to left column above action bar
- extract DiscussionMetaBar (counts + analytics + comment sort)
- render it on the left above the action bar, production-style
- reuse it in PostDiscussionPanel via showMetaBar (reader keeps it)
- widen comments sidebar to 400px and the card container to fit 768+400
- tighten gaps across the discovery columns and panel
