+++ packages/shared/src/components/post/discovery/PostFocusCard.tsx
@@ -24,6 +24,7 @@ import { combinedClicks } from '../../../lib/click';
 import { useFeature } from '../../GrowthBookProvider';
 import { feature } from '../../../lib/featureManagement';
 import { SourceStrip } from '../reader/SourceStrip';
+import { DiscussionMetaBar } from './DiscussionMetaBar';
 import { PostDiscussionPanel } from './PostDiscussionPanel';
 
 const PostCodeSnippets = dynamic(() =>
@@ -93,7 +94,7 @@ export const PostFocusCard = ({
       data-testid="post-focus-card"
     >
       <PostContainer className="relative laptop:shrink laptop:grow-0 laptop:basis-[768px] laptop:border-r-0">
-        <div className="flex min-w-0 flex-col gap-6 py-6 laptop:py-8">
+        <div className="flex min-w-0 flex-col gap-4 py-6 laptop:py-6">
           <div className="flex min-h-8 min-w-0 items-center gap-2">
             {post.source && (
               <SourceStrip
@@ -204,6 +205,7 @@ export const PostFocusCard = ({
             </div>
           )}
 
+          <DiscussionMetaBar post={post} />
           <PostActions
             post={post}
             postQueryKey={['post', post.id]}
@@ -214,10 +216,11 @@ export const PostFocusCard = ({
         </div>
       </PostContainer>
 
-      <aside className="flex min-h-0 min-w-0 shrink-0 flex-col border-t border-border-subtlest-tertiary bg-background-subtle laptop:sticky laptop:top-16 laptop:h-[calc(100vh-4rem)] laptop:max-h-[calc(100vh-4rem)] laptop:w-[340px] laptop:border-t-0 laptop:bg-background-default">
+      <aside className="flex min-h-0 min-w-0 shrink-0 flex-col border-t border-border-subtlest-tertiary bg-background-subtle laptop:sticky laptop:top-16 laptop:h-[calc(100vh-4rem)] laptop:max-h-[calc(100vh-4rem)] laptop:w-[400px] laptop:border-t-0 laptop:bg-background-default">
         <div className="flex min-h-0 w-full flex-col overflow-hidden rounded-24 border border-border-subtlest-tertiary bg-background-default shadow-2 laptop:h-full">
           <PostDiscussionPanel
             className="h-full p-4"
+            showMetaBar={false}
             onRegisterFocusComment={(fn) => {
               focusCommentRef.current = fn;
             }}
